msm: ipc: Remove incorrect use of smd_private.h
smd_private.h is an internal header file used to share private internal
details of the SMD driver to related components for the purposes of
debugging, segregating functionality, etc.  Nothing in smd_private.h
is intended to be exposed for client use, and including it in a file
should be a special exception, not a normal usecase.

The IPC framework is using smd_private.h as a means to access the SMSM
public API.  Substitute the SMSM public api headers for the uses of
smd_private.h.

Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>